+++ app/Http/Controllers/AdminController.php
@@ -12,16 +12,33 @@ use App\Models\Admin;
 use App\Models\Filter;
 use App\Models\Log;
 
+use App\Services\AdminService;
+
 use function PHPUnit\Framework\isEmpty;
 
 class AdminController extends Controller
 {
+    private $adminService;
+
+    public function __construct(AdminService $adminService) {
+        $this->adminService = $adminService;
+    }
+
     public function showAdminPanel() {
+        if (session()->has('products'))
+            $products = session('products');
+
+        if (session()->has('logs'))
+            $logs = session('logs');
+
+        if (session()->has('user'))
+            $user = session('user');
 
         return view('admin.admin-panel')->with([
             'success' => 'Welcome to Admin Panel.',
-            'products' => collect(),
-            'logs' => collect(),
+            'products' => $products ?? collect(),
+            'logs' => $logs ?? collect(),
+            'user' => $user ?? null,
         ]);
     }
 
@@ -32,7 +49,7 @@ class AdminController extends Controller
 
         $user = User::where('username', $request->username)->first();
         if (!$user) {
-            return back()->with('error', "The user doesn't exist.");
+            return back()->with('error', "The user doesn't exist.")->withInput();
         }
 
         $products = $user->products;
@@ -43,10 +60,31 @@ class AdminController extends Controller
 
         $logs = Log::where('causer_id', $user->id)->get();
 
-        return view('admin.admin-panel')->with([
+        return redirect()->route('admin-panel')->with([
             'products' => $products_coll,
             'logs' => $logs,
+            'user' => $user,
+        ])->withInput();
+    }
+
+    public function updateUser(Request $request, $id) {
+        $request->validate([
+            'username' => 'required|string|max:100',
+            'email' => 'required|email|max:100',
+            'password' => 'nullable|min:6|regex:/^(?=.*[0-9])(?=.*[\W_]).+$/',
         ]);
+
+        $user = User::find($id);
+        if ($user->username !== $request->username)
+            $this->adminService->changeUsername($request, $id);
+
+        if($user->email !== $request->email)
+            $this->adminService->changeEmail($request, $id);
+
+        if (isset($request->password))
+            $this->adminService->changePassword($request, $id);
+
+        return back()->with('success', 'You have successfully updated users data.');
     }
     
 }